product: Disable traced persistence by default
This change sets the default value of persist.traced.enable to 0,
disabling the system tracing daemon (traced) from persisting across
reboots.

While the tracing infrastructure is valuable for debugging, it is not
needed for the majority of users and can be safely disabled to
conserve system resources. Developers and users who require system
tracing can still enable it manually as needed.

Signed-off-by: Yehonatan5785 <yehonatan@mail.ru>