build: Disable compression for resource packaging
This change adds the -L 0 flag to disable compression when using
SOONG_ZIP for resource packaging. This applies to the command that
creates an archive from a file list.

Disabling compression significantly speeds up the packaging process
by eliminating the CPU overhead, at the cost of a slightly larger
file size. This is a beneficial trade-off for build performance
during development.

Signed-off-by: Yehonatan5785 <yehonatan@mail.ru>