build/make: Disable compression for target files package
This change adds the -L 0 flag to the SOONG_ZIP command when creating
the target files package. The -L 0 flag disables compression for all
entries added to the zip archive.

This can significantly speed up the build process during the
final packaging step, as it eliminates the CPU overhead of
compression, at the cost of a larger final file size.

Signed-off-by: Yehonatan5785 <yehonatan@mail.ru>